+++ packages/better_player/test/core/better_player_controller_advanced_test.dart
@@ -44,7 +44,7 @@ void main() {
 
       // Simulate error
       mock.value = mock.value.copyWith(errorDescription: 'Error');
-      controller.engineController!.notifyListeners();
+      mock.notifyListeners();
 
       await controller.retryDataSource();
       expect(controller.isEngineReady, true);